requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.0.3
django-cors-headers==2.2.0
django-heroku==0.3.1
django-tastypie==0.14.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.1
python-dateutil==2.7.2
python-mimeparse==1.6.0
pytz==2018.3
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
